trajectory_evaluator/requirements.txt
# Core dependencies
arize-phoenix>=4.29.0
arize-phoenix-evals>=0.13.0
openai>=1.0.0
httpx<0.28

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Async support
nest-asyncio>=1.5.0

# Utilities
python-dotenv>=1.0.0
